+++ Components/IRacing/Events/IRacingCommandSendRaceFlags.cs
@@ -1,6 +1,7 @@
-﻿using System.Collections.Generic;
+﻿using Slipstream.Shared;
+using System.Collections.Generic;
 
-namespace Slipstream.Shared.Events.IRacing
+namespace Slipstream.Components.IRacing.Events
 {
     public class IRacingCommandSendRaceFlags : IEvent
     {
