+++ Components/IRacing/Events/IRacingCommandSendWeatherInfo.cs
@@ -1,16 +1,17 @@
-﻿using System.Collections.Generic;
+﻿using Slipstream.Shared;
+using System.Collections.Generic;
 
-namespace Slipstream.Shared.Events.IRacing
+namespace Slipstream.Components.IRacing.Events
 {
-    public class IRacingCommandSendCarInfo : IEvent
+    public class IRacingCommandSendWeatherInfo : IEvent
     {
-        public string EventType => "IRacingCommandSendCarInfo";
+        public string EventType => "IRacingCommandSendWeatherInfo";
         public bool ExcludeFromTxrx => false;
         public ulong Uptime { get; set; }
 
         public override bool Equals(object obj)
         {
-            return obj is IRacingCommandSendCarInfo info &&
+            return obj is IRacingCommandSendWeatherInfo info &&
                    EventType == info.EventType &&
                    ExcludeFromTxrx == info.ExcludeFromTxrx;
         }
