+++ Components/IRacing/Events/IRacingRaceFlags.cs
@@ -1,6 +1,7 @@
+using Slipstream.Shared;
 using System.Collections.Generic;
 
-namespace Slipstream.Shared.Events.IRacing
+namespace Slipstream.Components.IRacing.Events
 {
     public class IRacingRaceFlags : IEvent
     {
