+++ Components/IRacing/Events/IRacingTesting.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Slipstream.Shared.Events.IRacing
+namespace Slipstream.Components.IRacing.Events
 {
     public class IRacingTesting : IIRacingSessionState
     {
