+++ Components/IRacing/Events/IRacingWarmup.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Slipstream.Shared.Events.IRacing
+namespace Slipstream.Components.IRacing.Events
 {
     public class IRacingWarmup : IIRacingSessionState
     {
