+++ Components/IRacing/Plugins/IRacingPlugin.cs
@@ -1,7 +1,7 @@
 ﻿using iRacingSDK;
 using Slipstream.Backend.Plugins;
+using Slipstream.Components.IRacing.Events;
 using Slipstream.Shared;
-using Slipstream.Shared.Events.IRacing;
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
