+++ Slipstream.csproj
@@ -166,13 +166,13 @@
     <Compile Include="Components\IComponentPluginCreationContext.cs" />
     <Compile Include="Components\IComponentPluginDependencies.cs" />
     <Compile Include="Components\IComponentRegistrationContext.cs" />
-    <Compile Include="Shared\Events\IRacing\IIRacingSessionState.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingCarPosition.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingWarmup.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingRace.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingTesting.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingQualify.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingPractice.cs" />
+    <Compile Include="Components\IRacing\Events\IIRacingSessionState.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingCarPosition.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingWarmup.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingRace.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingTesting.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingQualify.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingPractice.cs" />
     <Compile Include="Shared\Events\Twitch\TwitchRaided.cs" />
     <Compile Include="Shared\Events\Twitch\TwitchGiftedSubscription.cs" />
     <Compile Include="Shared\Events\Twitch\TwitchUserSubscribed.cs" />
@@ -216,12 +216,12 @@
     <Compile Include="Components\FileMonitor\Events\FileMonitorScanCompleted.cs" />
     <Compile Include="Shared\Events\Lua\LuaCommandDeduplicateEvents.cs" />
     <Compile Include="Shared\Events\Internal\InternalCommandPluginStates.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingCommandSendTrackInfo.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingCommandSendSessionState.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingCommandSendRaceFlags.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingCommandSendWeatherInfo.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingDriverIncident.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingCommandSendCarInfo.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingCommandSendTrackInfo.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingCommandSendSessionState.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingCommandSendRaceFlags.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingCommandSendWeatherInfo.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingDriverIncident.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingCommandSendCarInfo.cs" />
     <Compile Include="Shared\Events\Twitch\TwitchCommandSendWhisper.cs" />
     <Compile Include="Shared\Events\Twitch\TwitchReceivedWhisper.cs" />
     <Compile Include="Shared\Events\Twitch\TwitchReceivedMessage.cs" />
@@ -241,26 +241,26 @@
     <Compile Include="Shared\Factories\TwitchEventFactory.cs" />
     <Compile Include="Components\FileMonitor\EventHandler\FileMonitor.cs" />
     <Compile Include="Shared\ApplicationVersionService.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingPitstopReport.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingPitExit.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingPitEnter.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingCarCompletedLap.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingRaceFlags.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingCarInfo.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingPitstopReport.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingPitExit.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingPitEnter.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingCarCompletedLap.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingRaceFlags.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingCarInfo.cs" />
     <Compile Include="Components\IRacing\Plugins\IRacingPlugin.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingWeatherInfo.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingWeatherInfo.cs" />
     <Compile Include="Backend\Plugins\LuaPlugin.cs" />
     <Compile Include="Backend\Plugins\LuaManagerPlugin.cs" />
     <Compile Include="Components\FileMonitor\Plugins\FileMonitorPlugin.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingTrackInfo.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingTrackInfo.cs" />
     <Compile Include="Backend\PluginWorker.cs" />
     <Compile Include="Shared\EventHandlerController.cs" />
     <Compile Include="Components\FileMonitor\Events\FileMonitorFileChanged.cs" />
     <Compile Include="Components\FileMonitor\Events\FileMonitorFileDeleted.cs" />
     <Compile Include="Components\FileMonitor\Events\FileMonitorFileRenamed.cs" />
     <Compile Include="Components\FileMonitor\Events\FileMonitorFileCreated.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingDisconnected.cs" />
-    <Compile Include="Shared\Events\IRacing\IRacingConnected.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingDisconnected.cs" />
+    <Compile Include="Components\IRacing\Events\IRacingConnected.cs" />
     <Compile Include="Shared\Events\Internal\InternalCommandPluginUnregister.cs" />
     <Compile Include="Shared\Events\Internal\InternalCommandPluginRegister.cs" />
     <Compile Include="Shared\Events\Twitch\TwitchCommandSendMessage.cs" />
@@ -358,7 +358,9 @@
       <Name>iRacingSDK</Name>
     </ProjectReference>
   </ItemGroup>
-  <ItemGroup />
+  <ItemGroup>
+    <Folder Include="Shared\Events\IRacing\" />
+  </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="packages\KeraLua.1.2.11\build\net45\KeraLua.targets" Condition="Exists('packages\KeraLua.1.2.11\build\net45\KeraLua.targets')" />
   <Target Name="EnsureNuGetPackageBuildImports" BeforeTargets="PrepareForBuild">
